changed to manual serialization

+++ core/src/com/nkcoding/spacegame/simulation/communication/UpdateTransmission.java
@@ -1,16 +1,17 @@
 package com.nkcoding.spacegame.simulation.communication;
 
+import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 
 public class UpdateTransmission {
-    public final int updateID;
+    public final short updateID;
 
-    public UpdateTransmission(int updateID) {
+    public UpdateTransmission(short updateID) {
         this.updateID = updateID;
     }
 
     public void serialize(DataOutputStream outputStream) throws IOException {
-        outputStream.writeInt(updateID);
+        outputStream.writeShort(updateID);
     }
 }
